// Insurance Management Frontend/InsuranceManagementSystem/src/app/models/customer.ts
export interface Customer {

    customerId: string;
    prefix: string;
    firstName: string;
    middleName: string;
    lastName: string;
    email: string;
    contactNumber: string;
    alternateNumber: string;
    dateOfBirth: string;
    maritalStatus: string;
    gender: string;
    address: string;

}